--- src/helpers/isTimestamp.ts
@@ -1,3 +1,0 @@
-export function isTimestamp(timestamp: string) {
-    return timestamp.length === 13 && /^\d+$/.test(timestamp);
-}
