--- src/helpers/isUnixTimestamp.ts
@@ -1,3 +1,0 @@
-export function isUnixTimestamp(timestamp: string): boolean {
-    return /^\d+$/.test(timestamp) && timestamp.length === 10;
-}
